pyproject.toml
[project]
name = "sgh-api"
version = "0.1.0"
description = "A RESTful API for accessing admission announcements and related programs from educational institutions."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
]
